requirements.txt
fpdf==1.7.2
Pillow==6.2.1
selenium==3.141.0
urllib3==1.25.7
uuid==1.30
